+++ visualization/app/codeCharta/state/effects/unfocusNodesOnLoadingMap.effect.spec.ts
@@ -2,6 +2,7 @@ import { ApplicationInitStatus } from "@angular/core"
 import { TestBed } from "@angular/core/testing"
 import { Action } from "redux"
 import { Subject } from "rxjs"
+import { mocked } from "ts-jest/utils"
 
 import { EffectsModule } from "../angular-redux/effects/effects.module"
 import { setIsLoadingFile } from "../store/appSettings/isLoadingFile/isLoadingFile.actions"
@@ -9,9 +10,14 @@ import { unfocusAllNodes } from "../store/dynamicSettings/focusedNodePath/focuse
 import { Store } from "../store/store"
 import { UnfocusNodesOnLoadingMapEffect } from "./unfocusNodesOnLoadingMap.effect"
 
+jest.mock("../store/store", () => ({
+	Store: {}
+}))
+const MockedStore = mocked(Store)
+
 describe("UnfocusNodesOnLoadingMapEffect", () => {
 	beforeEach(async () => {
-		Store["initialize"]()
+		MockedStore.dispatch = jest.fn()
 
 		EffectsModule.actions$ = new Subject<Action>()
 		TestBed.configureTestingModule({
@@ -24,15 +30,18 @@ describe("UnfocusNodesOnLoadingMapEffect", () => {
 		EffectsModule.actions$.complete()
 	})
 
-	it("should do nothing on a not relevant action", () => {
-		const dispatchSpy = jest.spyOn(Store.store, "dispatch")
+	it("should ignore a not relevant action", () => {
 		EffectsModule.actions$.next({ type: "whatever" })
-		expect(dispatchSpy).toHaveBeenCalledTimes(0)
+		expect(MockedStore.dispatch).not.toHaveBeenCalled()
 	})
 
-	it("should fire an unfocusAllNodes action on a setIsLoadingFile action", () => {
-		const dispatchSpy = jest.spyOn(Store.store, "dispatch")
+	it("should fire an unfocusAllNodes action after setIsLoadingFile to true", () => {
 		EffectsModule.actions$.next(setIsLoadingFile(true))
-		expect(dispatchSpy).toHaveBeenCalledWith(unfocusAllNodes())
+		expect(MockedStore.dispatch).toHaveBeenCalledWith(unfocusAllNodes())
+	})
+
+	it("should ignore setIsLoadingFile to false", () => {
+		EffectsModule.actions$.next(setIsLoadingFile(false))
+		expect(MockedStore.dispatch).not.toHaveBeenCalled()
 	})
 })