+++ visualization/app/codeCharta/state/effects/unfocusNodesOnLoadingMap.effect.ts
@@ -1,9 +1,9 @@
 import { Inject } from "@angular/core"
 import { filter, map } from "rxjs"
-import { isActionOfType } from "../../util/reduxHelper"
+import { isAction } from "../../util/reduxHelper"
 import { createEffect } from "../angular-redux/effects/createEffect"
 import { Actions, ActionsToken } from "../angular-redux/effects/effects.module"
-import { IsLoadingFileActions } from "../store/appSettings/isLoadingFile/isLoadingFile.actions"
+import { IsLoadingFileActions, SetIsLoadingFileAction } from "../store/appSettings/isLoadingFile/isLoadingFile.actions"
 import { unfocusAllNodes } from "../store/dynamicSettings/focusedNodePath/focusedNodePath.actions"
 
 export class UnfocusNodesOnLoadingMapEffect {
@@ -11,7 +11,7 @@ export class UnfocusNodesOnLoadingMapEffect {
 
 	unfocusNodesOnLoadingMap$ = createEffect(() => {
 		return this.actions$.pipe(
-			filter(action => isActionOfType(action.type, IsLoadingFileActions)),
+			filter(action => isAction<SetIsLoadingFileAction>(action, IsLoadingFileActions) && action.payload === true),
 			map(() => unfocusAllNodes())
 		)
 	})